rename files and routes to better match the schema
