+++ web/src/app/admin/admin-outgoing-messages/DebugMessageDetails.tsx
@@ -20,7 +20,7 @@ interface Props {
   log?: DebugMessage | null
 }
 
-export default function OutgoingLogDetails(props: Props): JSX.Element {
+export default function DebugMessageDetails(props: Props): JSX.Element {
   const { open, onClose, log } = props
 
   return (
